Rename TASK 2 Password generator to TASK 2 Password Generator
As part of my virtual internship at CodSoft, I had the opportunity to work on an exciting project: creating a Password Generator application using Python. This tool is designed to help users generate strong and random passwords with ease.
Project Highlights:
User Input: Prompts users to specify the desired password length.
Generate Password: Uses a combination of random characters to generate a password of the specified length.
Display the Password: Prints the generated password on the screen.
Working on this project has been a fantastic learning experience, allowing me to apply my Python skills to enhance online security. Proud to be part of the Codsoft team and contribute to meaningful projects!
hashtag#codsoft hashtag#internship